Add dark mode, handle Chrome enable-force-dark flag
Adds the ability to switch themes with a dropdown, and automatically detects the user's preferred theme on startup.

Also handles the chrome flag `#enable-force-dark`, which seems to cause issues with MathJax's fraction bar and square root line if the body color is set to `#fff`, but the issue goes away if they are set to `#cdcdcd` (but not `#cecece`.