core: handle degree/minutes/seconds parsing
I still need to add the LaTeX handling of this

NOTE: originally written on 2024-05-30,
but on 2026-06-05 updated to latest version...
but had to change the float pattern from
[0-9]+ to [0-9]*
so that leading zero could be dropped, e.g.
".6" can be parsed instead of just "0.6".

But I need to see if this breaks anything else
introduced in this commit.