android: update js_display.html to use async MathJax API
I'm not sure why the other one stopped working. I'm pretty sure
it still works on the published app, and I am somewhat sure that
it used to work in a modern browser. I didn't think it had to do
with adding the doctype to avoid quirks mode, but I didn't
investigate too much.