android: increase version code to 9 and name to "1.0.5"
These commits are out of order, the changes I made in the
commit to target SDK level 34 and increase to version
code 8 and name "1.0.4" happened a long time ago and were
published to Google Play.

This new version will include all changes that I've made since
publishing the source to github.